upstream merge

+++ neutron_taas/services/taas/taas_plugin.py
@@ -18,9 +18,9 @@ from neutron.db import servicetype_db as st_db
 from neutron.extensions import portbindings
 from neutron.services import provider_configuration as pconf
 from neutron.services import service_base
-from neutron_lib.callbacks import events
-from neutron_lib.callbacks import registry
-from neutron_lib.callbacks import resources
+from neutron.callbacks import events
+from neutron.callbacks import registry
+from neutron.callbacks import resources
 from neutron_lib import constants
 from neutron_lib import exceptions as n_exc
 
@@ -59,7 +59,7 @@ class TaasPlugin(taas_db.Taas_db_Mixin):
 
         registry.subscribe(self.handle_delete_port,
                            resources.PORT,
-                           events.PRECOMMIT_DELETE)
+                           events.BEFORE_DELETE)
 
         return
 
@@ -263,20 +263,31 @@ class TaasPlugin(taas_db.Taas_db_Mixin):
                               "tap_flow: %s", id)
 
     def handle_delete_port(self, resource, event, trigger, context, **kwargs):
-        deleted_port = kwargs['port']
-        LOG.info("TaaS: Handle Delete Port: %s", deleted_port['id'])
+        deleted_port_id = kwargs['port_id']
+        if not deleted_port_id:
+            LOG.error("TaaS: Handle Delete Port: Invalid port_id received")
+            return
 
-        # Get list of configured tap-services
-        t_s_collection = self.get_tap_services(
-            context,
-            filters={'port_id': [deleted_port['id']]}, fields=['id'])
+        LOG.info("TaaS: Handle Delete Port: %s", deleted_port_id)
 
-        for t_s in t_s_collection:
-            self.delete_tap_service(context, t_s['id'])
+        with context.session.begin(subtransactions=True):
+            # Get list of configured tap-services
+            t_s_collection = self.get_tap_services(
+                context,
+                filters={'port_id': [deleted_port_id]}, fields=['id'])
 
-        t_f_collection = self.get_tap_flows(
-            context,
-            filters={'source_port': [deleted_port['id']]}, fields=['id'])
+            for t_s in t_s_collection:
+                try:
+                    self.delete_tap_service(context, t_s['id'])
+                except taas_ex.TapServiceNotFound:
+                    LOG.debug("Not found tap_service: %s", t_s['id'])
 
-        for t_f in t_f_collection:
-            self.delete_tap_flow(context, t_f['id'])
+            t_f_collection = self.get_tap_flows(
+                context,
+                filters={'source_port': [deleted_port_id]}, fields=['id'])
+
+            for t_f in t_f_collection:
+                try:
+                    self.delete_tap_flow(context, t_f['id'])
+                except taas_ex.TapFlowNotFound:
+                    LOG.debug("Not found tap_flow: %s", t_f['id'])
